+++ BNA_KC_Vehicles/Light/Config.cpp
@@ -1,5 +1,4 @@
 #include "CfgPatches.hpp"
-#include "CfgCloudlets.hpp"
 
 
 class CfgWeapons
